Restructured store. Review fixes
