+++ src/store/profile/index.js
@@ -1,2 +1,3 @@
 export * from './actions';
 export * from './middlewares';
+export * from './reducers';
